Some bug fixes and testing
Need to make sure we are getting the correct id for all types of targets (spawns and controller).

Also may very well need to change the format for saving everything in the world to being ids instead of the object. Kind of wonder if that's why I keep losing control of my workers.